+++ tests/Utils/Resolvers/Foo.php
@@ -6,6 +6,11 @@ use Illuminate\Support\Arr;
 
 class Foo
 {
+    public function __invoke(): string
+    {
+        return 'foo.baz';
+    }
+
     public function bar(): string
     {
         return 'foo.bar';