+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.1.0...master)
+## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.1.1...master)
+
+## [4.1.1](https://github.com/nuwave/lighthouse/compare/v4.1.0...v4.1.1)
+
+### Fixed
+
+- Unbox laravel-enum inputs when using the builder directives https://github.com/nuwave/lighthouse/pull/927
 
 ## [4.1.0](https://github.com/nuwave/lighthouse/compare/v4.0.0...v4.1.0)